Replace `LeafNodeRef` with `LeafIndex` everywhere (#332)
* Remove `LeafNodeRef` and replace with `LeafIndex`

* Implement remove filter

* Make RustFmt happy

* Fix rust fmt ... again

+++ src/group/epoch.rs
@@ -267,7 +267,7 @@ pub mod test_utils {
     pub(crate) fn get_test_epoch(cipher_suite: CipherSuite) -> Epoch {
         let secret_tree = get_test_tree(
             cipher_suite,
-            vec![0 as u8; Hkdf::from(cipher_suite.kdf_type()).extract_size()],
+            vec![0_u8; Hkdf::from(cipher_suite.kdf_type()).extract_size()],
             2,
         );
